Added new Event serialization tests and integration tests for Metadata Transactions.
